pyproject.toml
[project]
name = "hackatontibd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mysql-connector-python>=9.4.0",
    "mysqlclient",
    "polars>=1.33.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
